Swaps attribute and content order in node creation.
Prior, when calling `lmth.$element`, the first argument could be a
string `content`, and if there was a second string argument, that
would be made the `attributes`. Now, if there is a first string
argument, it is made the `attributes`, and the second string argument
is made the `content`:

    `lmth.div('content', '#top-nav.hero.col-6')` =>
    `lmth.div('#top-nav.hero.col-6', 'content')`

To create a node with no attributes, yet with content:

    `lmth.$element(null, 'content')`

The rationale is that nodes with only ids and classes are used far
more often than nodes with textual content, and thus should be easier
to create.

This is a breaking change.