Defers attribute value to-string conversion to rendering.
Prior, all attribute values were converted to strings upon element
creation. This was necessary because rendering to a string requires
that all attribute values be strings. Now, this conversion is only
done upon actual string rendering, thus saving some cycles up front.
Furthermore, if you never render a string, or render to a DOM, you no
longer pay for that initial 'convert-all-attribute-values-to-strings'
loop (because the DOM's `element.setAttribute` handles that for us
instead).

N.B.: This is only generally relevant for numeric and boolean
primitives.

This is (probably) a breaking change.